include authority data for performance graphs
Also separates datatable and graph behaviors into concerns for performance_presenter.
