+++ app/services/qa_server/performance_graph_service.rb
@@ -18,11 +18,13 @@ module QaServer
       #     ...,
       #     23: { hour: 'NOW', stats: { load_avg_ms: 12.3, normalization_avg_ms: 4.2, full_request_avg_ms: 16.5, etc. }}
       #   }
-      def average_last_24_hours
+      def average_last_24_hours(auth_name = nil)
         start_hour = Time.now.getlocal.beginning_of_hour - 23.hours
         avgs = {}
         0.upto(23).each do |idx|
-          records = performance_data_class.where(dt_stamp: start_hour..start_hour.end_of_hour)
+          where_clause = { dt_stamp: start_hour..start_hour.end_of_hour }
+          where_clause[:authority] = auth_name unless auth_name.nil?
+          records = performance_data_class.where(where_clause)
           stats = stats_calculator_class.calculate_stats(records)
           data = {}
           data[BY_HOUR] = performance_by_hour_label(idx, start_hour)
@@ -31,16 +33,6 @@ module QaServer
           start_hour += 1.hour
         end
         avgs
-      end
-
-      def performance_by_hour_label(idx, start_hour)
-        if idx == 23
-          I18n.t('qa_server.monitor_status.performance.now')
-        elsif ((idx + 1) % 2).zero?
-          (start_hour.hour * 100).to_s
-        else
-          ""
-        end
       end
 
       # Get daily average for the past 30 days.
@@ -52,11 +44,13 @@ module QaServer
       #     ...,
       #     29: { day: 'TODAY', stats: { load_avg_ms: 12.3, normalization_avg_ms: 4.2, full_request_avg_ms: 16.5, etc. }}
       #   }
-      def average_last_30_days
+      def average_last_30_days(auth_name = nil)
         start_day = Time.now.getlocal.beginning_of_day - 29.days
         avgs = {}
         0.upto(29).each do |idx|
-          records = performance_data_class.where(dt_stamp: start_day..start_day.end_of_day)
+          where_clause = { dt_stamp: start_day..start_day.end_of_day }
+          where_clause[:authority] = auth_name unless auth_name.nil?
+          records = performance_data_class.where(where_clause)
           stats = stats_calculator_class.calculate_stats(records)
           data = {}
           data[BY_DAY] = performance_by_day_label(idx, start_day)
@@ -65,16 +59,6 @@ module QaServer
           start_day += 1.day
         end
         avgs
-      end
-
-      def performance_by_day_label(idx, start_day)
-        if idx == 29
-          I18n.t('qa_server.monitor_status.performance.today')
-        elsif ((idx + 1) % 5).zero?
-          start_day.strftime("%m-%d")
-        else
-          ""
-        end
       end
 
       # Get daily average for the past 12 months.
@@ -86,11 +70,13 @@ module QaServer
       #     ...,
       #     11: { month: '08-2019', stats: { load_avg_ms: 12.3, normalization_avg_ms: 4.2, full_request_avg_ms: 16.5, etc. }}
       #   }
-      def average_last_12_months
+      def average_last_12_months(auth_name = nil)
         start_month = Time.now.getlocal.beginning_of_month - 11.months
         avgs = {}
         0.upto(11).each do |idx|
-          records = performance_data_class.where(dt_stamp: start_month..start_month.end_of_month)
+          where_clause = { dt_stamp: start_month..start_month.end_of_month }
+          where_clause[:authority] = auth_name unless auth_name.nil?
+          records = performance_data_class.where(where_clause)
           stats = stats_calculator_class.calculate_stats(records)
           data = {}
           data[BY_MONTH] = start_month.strftime("%m-%Y")
@@ -100,6 +86,28 @@ module QaServer
         end
         avgs
       end
+
+      private
+
+        def performance_by_hour_label(idx, start_hour)
+          if idx == 23
+            I18n.t('qa_server.monitor_status.performance.now')
+          elsif ((idx + 1) % 2).zero?
+            (start_hour.hour * 100).to_s
+          else
+            ""
+          end
+        end
+
+        def performance_by_day_label(idx, start_day)
+          if idx == 29
+            I18n.t('qa_server.monitor_status.performance.today')
+          elsif ((idx + 1) % 5).zero?
+            start_day.strftime("%m-%d")
+          else
+            ""
+          end
+        end
     end
   end
 end
